Add current user validation to post update and delete endpoints

+++ routers/posts.py
@@ -55,7 +55,11 @@ async def create_post(post: PostCreate, current_user: CurrentUser, db: Annotated
 
 
 @router.put("/{post_id}", response_model=PostResponse)
-def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+def update_post_full(post_id: int,
+                     post_data: PostCreate,
+                     current_user: CurrentUser,
+                     db: Annotated[AsyncSession, Depends(get_db)]):
+
     results = db.execute(select(models.Post).where(models.Post.id == post_id))
     post = results.scalars().first()
     print(f"This is the post we are printing: {post}")
@@ -63,15 +67,22 @@ def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSes
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post Not Fount")
 
-    if post_data.user_id != post.user_id:
-        results = db.execute(select(models.User).where(models.User.id == post_data.user_id))
-        user = results.scalars().first()
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User Not Found")
+    """This is commented out bcz of the CurrentUser added and this will be handled there"""
+    # if post_data.user_id != post.user_id:
+    #     results = db.execute(select(models.User).where(models.User.id == post_data.user_id))
+    #     user = results.scalars().first()
+    #     if not user:
+    #         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User Not Found")
+    if post.user_id != current_user.id:
+        # Why 403 instead of 401 ?
+        # 403: user is authenticated, but, you don't have permission to update the post
+        # 401: un-authorized user - invalid token for the current user
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to update this post")
 
     post.title = post_data.title
     post.content = post_data.content
-    post.used_id = post_data.user_id
+    # post.used_id = post_data.user_id
 
     db.commit()
     db.refresh(post, attribute_names=["author"])
@@ -80,13 +91,24 @@ def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSes
 
 
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id: int,
+                              post_data: PostUpdate,
+                              current_user: CurrentUser,
+                              db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     print(f"This is the post we are printing: {post}")
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post Not Fount")
+
+    if post.user_id != current_user.id:
+        # Why 403 instead of 401 ?
+        # 403: user is authenticated, but, you don't have permission to update the post
+        # 401: un-authorized user - invalid token for the current user
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to update this post")
+
 
     update_data = post_data.model_dump(exclude_unset=True)
     print(f"Update data is: {update_data}")
@@ -101,12 +123,21 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int,
+                      current_user: CurrentUser,
+                      db: Annotated[AsyncSession, Depends(get_db)]):
     results = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = results.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post Not Fount")
+
+    if post.user_id != current_user.id:
+        # Why 403 instead of 401 ?
+        # 403: user is authenticated, but, you don't have permission to update the post
+        # 401: un-authorized user - invalid token for the current user
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to delete this post")
 
     await db.delete(post)
     await db.commit()
